\name{calcCrabFisheryQuarterFromMonth}
\alias{calcCrabFisheryQuarterFromMonth}
\title{Convert month(s) to quarters by crab fishery year (starting July 1).}
\usage{
calcCrabFisheryQuarterFromMonth(x)
}
\arguments{
  \item{x}{- vector of months to convert}
}
\value{
vector of crab fishery quarters
}
\description{
Function that converts a vector of months to a vector of
quarters by crab fishery year (i.e., starting July 1).
}
\details{

}

