% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parseStatAreas}
\alias{parseStatAreas}
\title{Convert statareas to centered lat/lons.}
\usage{
parseStatAreas(x)
}
\arguments{
\item{x}{- vector of statareas to convert}
}
\value{
list with two vector elements: 'lon' and 'lat'
}
\description{
Function that converts a vector of statareas to a
list of lat, lon vectors corresponding to the centers of the statareas.
}
\details{

}

